feat:Enhances user security and data integrity
Increases password length and sets default value for CreatedAt fields.
Additionally, it ensures that only Admin roles can create users and games.

The password length is increased to 100 characters.
The CreatedAt field is configured to default to the current timestamp in the database.
Modifies Game and User services to validate user role and ensure that only administrators can create Games and Users.